+++ CHANGELOG.md
@@ -7,9 +7,8 @@
   to call `rm -rf`.
 - Deprecate `--username` flag on and username prompt on `buf registry login`. A username is no longer
   required to log in.
-- Add `--list` and `--describe` flags to `buf curl`, which trigger the command to list
-  known services in the RPC schema or to describe an element (such as a method, message,
-  etc), instead of invoking an RPC method.
+- Add `--list-services` and `--list-methods` flags to `buf curl`, which trigger the command to list
+  known services or methods in the RPC schema, instead of invoking an RPC method.
 
 ## [v1.34.0] - 2024-06-21